common method for Retrieves the appropriate Azure credential
