+++ infra/scripts/process_data_scripts.sh
@@ -12,6 +12,7 @@ echo "Script Started"
 
 curl --output "04_cu_process_data_new_data.py" ${baseUrl}"infra/scripts/index_scripts/04_cu_process_data_new_data.py"
 curl --output "content_understanding_client.py" ${baseUrl}"infra/scripts/index_scripts/content_understanding_client.py"
+curl --output "azure_credential_utils.py" ${baseUrl}"infra/scripts/index_scripts/azure_credential_utils.py"
 curl --output "ckm-analyzer_config_text.json" ${baseUrl}"infra/data/ckm-analyzer_config_text.json"
 curl --output "ckm-analyzer_config_audio.json" ${baseUrl}"infra/data/ckm-analyzer_config_audio.json"
